+++ Task_1/src/Locomotive.java
@@ -2,7 +2,7 @@ public class Locomotive extends Train {
 	
 	private int coefficientOfPerformance;
 
-	Locomotive(String name, double currentSpeed, double weight, 
+	Locomotive(String name, int currentSpeed, double weight, 
 				int countOfPassengers, int countOfWheels) {
 		super(name, currentSpeed, weight, countOfPassengers, countOfWheels);
 	}
